Define the public API of the Library
The Library will have one public class which allows users to interact with the library.
The user will be able to send directly to a device, to send indirectly, to disconnect connections and
many interactions more.
Most importantly, the user has to register a listener for the events defined in ICommunicationEvent.
The user also has to register a ILogFragment to view log messages.
The BluetoothConnection-class wraps the connection via bluetooth and adds some data to it.